+++ src-docs/src/views/header/header_updates.js
@@ -8,7 +8,6 @@ import {
   EuiFlyoutBody,
   EuiFlyoutHeader,
   EuiTitle,
-  EuiNotificationBadge,
   EuiLink,
   EuiFlyoutFooter,
   EuiFlexGroup,
@@ -121,15 +120,12 @@ export default class extends Component {
         aria-controls="headerNewsFeed"
         aria-expanded={this.state.isFlyoutVisible}
         aria-haspopup="true"
-        aria-label="News feed"
-        onClick={this.showFlyout}>
+        aria-label={`News feed: ${
+          this.state.showBadge ? 'Updates available' : 'No updates'
+        }`}
+        onClick={this.showFlyout}
+        notification={this.state.showBadge && '•'}>
         <EuiIcon type="email" size="m" />
-
-        {this.state.showBadge ? (
-          <EuiNotificationBadge className="euiHeaderNotification">
-            &#9642;
-          </EuiNotificationBadge>
-        ) : null}
       </EuiHeaderSectionItemButton>
     );