+++ src/components/code_editor/code_editor.tsx
@@ -25,7 +25,9 @@ type SupportedAriaAttribute =
   | 'aria-describedby';
 type SupportedAriaAttributes = Pick<AriaAttributes, SupportedAriaAttribute>;
 
-export interface EuiCodeEditorProps extends SupportedAriaAttributes {
+export interface EuiCodeEditorProps
+  extends SupportedAriaAttributes,
+    Omit<IAceEditorProps, 'mode'> {
   width?: string;
   height?: string;
   onBlur?: IAceEditorProps['onBlur'];
@@ -34,6 +36,11 @@ export interface EuiCodeEditorProps extends SupportedAriaAttributes {
   setOptions: IAceEditorProps['setOptions'];
   cursorStart?: number;
   'data-test-subj'?: string;
+  /**
+   * Select the `brace` theme
+   * The matching theme file must also be imported from `brace` (e.g., `import 'brace/theme/github';`)
+   */
+  theme?: IAceEditorProps['theme'];
 
   /**
    * Use string for a built-in mode or object for a custom mode
@@ -169,6 +176,7 @@ export class EuiCodeEditor extends Component<
       cursorStart,
       mode = DEFAULT_MODE,
       'data-test-subj': dataTestSubj = 'codeEditorContainer',
+      theme = 'github',
       ...rest
     } = this.props;
 
@@ -253,6 +261,7 @@ export class EuiCodeEditor extends Component<
           // prior to dynamically setting a custom mode (https://github.com/elastic/eui/pull/2616)
           mode={this.isCustomMode() ? DEFAULT_MODE : (mode as string)} // https://github.com/securingsincity/react-ace/pull/771
           name={this.idGenerator()}
+          theme={theme}
           ref={this.aceEditorRef}
           width={width}
           height={height}