+++ src/components/index.js
@@ -4,9 +4,18 @@ export { EuiAspectRatio } from './aspect_ratio';
 
 export { EuiAvatar } from './avatar';
 
-export { EuiKeyboardAccessible, EuiScreenReaderOnly } from './accessibility';
+export {
+  EuiKeyboardAccessible,
+  EuiScreenReaderOnly,
+  EuiSkipLink,
+} from './accessibility';
 
-export { EuiBadge, EuiBetaBadge, EuiNotificationBadge } from './badge';
+export {
+  EuiBadge,
+  EuiBetaBadge,
+  EuiNotificationBadge,
+  EuiBadgeGroup,
+} from './badge';
 
 export { EuiBottomBar } from './bottom_bar';
 
@@ -177,6 +186,8 @@ export { EuiLink } from './link';
 
 export { EuiListGroup, EuiListGroupItem } from './list_group';
 
+export { EuiMark } from './mark';
+
 export {
   EUI_MODAL_CANCEL_BUTTON,
   EUI_MODAL_CONFIRM_BUTTON,
@@ -230,7 +241,10 @@ export { EuiProgress } from './progress';
 
 export { EuiTreeView } from './tree_view';
 
-export { EuiResizeObserver } from './observer/resize_observer';
+export {
+  EuiResizeObserver,
+  useResizeObserver,
+} from './observer/resize_observer';
 
 export { EuiSearchBar, Query, Ast } from './search_bar';